+++ app/javascript/applications/HomebrewsApp/requests/copyDaggerheartRecipe.jsx
@@ -1,0 +1,8 @@
+import { apiRequest, options } from '../helpers';
+
+export const copyDaggerheartRecipe = async (accessToken, id) => {
+  return await apiRequest({
+    url: `/homebrews/daggerheart/recipes/${id}/copy.json`,
+    options: options('POST', accessToken)
+  });
+}