+++ config/routes.rb
@@ -177,7 +177,12 @@ Rails.application.routes.draw do
       resources :domains, only: %i[index show create update destroy]
       resources :transformations, only: %i[index show create update destroy]
       resources :feats, only: %i[index create destroy]
-      resources :items, only: %i[index show create update destroy]
+      resources :items, only: %i[index show create update destroy] do
+        post :copy, on: :member
+      end
+      resources :recipes, only: %i[index create destroy] do
+        post :copy, on: :member
+      end
       resources :books, only: %i[index create update destroy] do
         resource :content, only: %i[create], module: :books
       end